Recreating Algebra class
