+++ parser/interpreter.py
@@ -1,6 +1,3 @@
-
-
-
 class Interpreter:
     def __init__(self, tree):
         self.tree = tree
